Initial OpenID integration
Added Facebook connect button to the end of the user login page. Modified the safacebook module to not force FB login to be logged in, and set a session variable if the user logged in via Facebook, so the correct logout link could be created. Added some styling for the user page.