Google Federated Open ID integration
Added lightopenid library and php files. The googleauth one is a modification of the example-google one. Putting this logic in a module was causing some strange issues of not logging the user in correctly, so for now, I'm using the modified php file and redirecting to a module for logging in the user.
Also added the google login button.