Custom openid javascript needed
I hate to do this this way, but we needed to remove the username and password fields on the login page. They wanted only outside ways like FB and Google to be the ways to login. However, Openid operates on those fields via Javascript, and I had to slightly alter it. Should come up with a way to do this better.